+++ frontend/front/src/app/page.tsx
@@ -1,7 +1,7 @@
 import Homee from "./views/Home";
 export default function Home() {
   return (
-    <div className="bg-[#b9362d]">
+    <div className="">
       <Homee />
     </div>
   );
